% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial.R
\name{$,ProbitSpatial-method}
\alias{$,ProbitSpatial-method}
\title{Extract from ProbitSpatial class.}
\usage{
\S4method{$}{ProbitSpatial}(x, name)
}
\arguments{
\item{x}{an object of class \code{ProbitSpatial}.}

\item{name}{of the slot.}
}
\value{
The content of the slot.
}
\description{
extract a slot from \code{ProbitSpatial} class object.
}
