% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial.R
\name{coef.ProbitSpatial}
\alias{coef.ProbitSpatial}
\title{Estimated coefficients of a spatial probit model.}
\usage{
\method{coef}{ProbitSpatial}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ProbitSpatial}.}

\item{...}{ignored}
}
\value{
It returns the value of the estimated parameters.
}
\description{
Returns the coefficients estimated by a \code{ProbitSpatial} model.
}
