% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial_R.R
\name{coef.SpatialProbit}
\alias{coef.SpatialProbit}
\title{Estimated coefficients of a spatial probit model.}
\usage{
\method{coef}{SpatialProbit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{SpatialProbit}.}

\item{...}{ignored}
}
\value{
It returns the value of the estimated parameters.
}
\description{
Returns the coefficients estimated by a \code{SpatialProbit} model.
}

