% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial_R.R
\name{conditional_SAR_UC}
\alias{conditional_SAR_UC}
\title{Conditional SAR UC.}
\usage{
conditional_SAR_UC(myenv)
}
\arguments{
\item{myenv}{an \code{environment}.}
}
\value{
the log-likelihood and the estimated parameters.
}
\description{
Performs conditional estimation of SAR model with variance-covariance matrix
}
\details{
We discourage the direct use of this function.
}

