% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial.R
\name{conditional_SAR_UP}
\alias{conditional_SAR_UP}
\title{Conditional SAR UP.}
\usage{
conditional_SAR_UP(myenv)
}
\arguments{
\item{myenv}{an \code{environment}.}
}
\value{
the log-likelihood and the estimated parameters.
}
\description{
Performs conditional estimation of SAR model with precision matrix.
}
\details{
We discourage the direct use of this function.
}
