% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial.R
\name{conditional_SARAR_UC}
\alias{conditional_SARAR_UC}
\title{Conditional SARAR UC.}
\usage{
conditional_SARAR_UC(myenv)
}
\arguments{
\item{myenv}{an \code{environment}.}
}
\value{
the log-likelihood and the estimated parameters.
}
\description{
Performs conditional estimation of SARAR model with variance-covariance 
matrix.
}
\details{
We discourage the direct use of this function.
}
