% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial.R
\name{conditional_SEM_UC}
\alias{conditional_SEM_UC}
\title{Conditional SEM UC.}
\usage{
conditional_SEM_UC(myenv)
}
\arguments{
\item{myenv}{an \code{environment}.}
}
\value{
the log-likelihood and the estimated parameters.
}
\description{
Performs conditional estimation of SEM model with variance-covariance matrix.
}
\details{
We discourage the direct use of this function.
}
