% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial_R.R
\name{conditional_SEM_UP}
\alias{conditional_SEM_UP}
\title{Conditional SEM UP.}
\usage{
conditional_SEM_UP(myenv)
}
\arguments{
\item{myenv}{an \code{environment}.}
}
\value{
the log-likelihood and the estimated parameters.
}
\description{
Performs conditional estimation of SEM model with precision matrix
}
\details{
We discourage the direct use of this function.
}

