% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial.R
\name{fitted.ProbitSpatial}
\alias{fitted.ProbitSpatial}
\title{Extract spatial probit model fitted values.}
\usage{
\method{fitted}{ProbitSpatial}(object, type = c("link", "response", "binary"), cut = 0.5, ...)
}
\arguments{
\item{object}{an object of class \code{ProbitSpatial}.}

\item{type}{the type of output: 
\describe{
  \item{\code{"link"}}{the value of the latent variable. Default.}
  \item{\code{"response"}}{probability.}
  \item{\code{"binary"}}{binary 0/1 output.}
}}

\item{cut}{the threshold probability for the \code{"binary"} type. 
Default is 0.5.}

\item{...}{ignored}
}
\value{
Returns the vector of fitted values of the \code{ProbitSpatial} model
}
\description{
Extract the fitted values of a \code{ProbitSpatial} model.
}
