% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial_R.R
\name{fitted.SpatialProbit}
\alias{fitted.SpatialProbit}
\title{Extract spatial probit model fitted values.}
\usage{
\method{fitted}{SpatialProbit}(object, type = "link", cut = 0.5, ...)
}
\arguments{
\item{object}{an object of class \code{SpatialProbit}.}

\item{type}{the type of prediction: 
\describe{
  \item{\code{"link"}}{the value of the latent variable. Default.}
  \item{\code{"response"}}{probability.}
  \item{\code{"binary"}}{binary 0/1 output.}
}}

\item{cut}{the threshold probability for the \code{"binary"} type. 
Default is 0.5.}

\item{...}{ignored}
}
\value{
Returns the vector of fitted values of the \code{SpatialProbit} model
}
\description{
Extract the fitted values of a \code{SpatialProbit} model.
}

