% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial.R
\name{names.ProbitSpatial}
\alias{names.ProbitSpatial}
\title{Extract names of ProbitSpatial class.}
\usage{
\method{names}{ProbitSpatial}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ProbitSpatial}.}

\item{...}{ignored}
}
\value{
Returns the names of the \code{ProbitSpatial} object.
}
\description{
Extract names of ProbitSpatial class.
}
