% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial_R.R
\name{names.SpatialProbit}
\alias{names.SpatialProbit}
\title{Extract names of SpatialProbit class.}
\usage{
\method{names}{SpatialProbit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{SpatialProbit}.}

\item{...}{ignored}
}
\value{
Returns the names of the \code{SpatialProbit} object.
}
\description{
Extract names of SpatialProbit class.
}

