% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial_R.R
\name{predict.SpatialProbit}
\alias{predict.SpatialProbit}
\title{Spatial probit model predictions.}
\usage{
\method{predict}{SpatialProbit}(object, X, type = "link", cut = 0.5, ...)
}
\arguments{
\item{object}{an object of class \code{SpatialProbit}.}

\item{X}{a martix of explanatory variables.}

\item{type}{the type of prediction: 
\describe{
  \item{\code{"link"}}{the value of the latent variable. Default}
  \item{\code{"response"}}{probability.}
  \item{\code{"binary"}}{binary 0/1 output.}
}}

\item{cut}{the threshold probability for the \code{"binary"} type. 
Default is 0.5.}

\item{...}{ignored}
}
\value{
Returns a vector of predicted values for the set \code{X} of 
	covariates
}
\description{
Predicts of a \code{SpatialProbit} model on a set \code{X} of covariates
}

