% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial.R
\name{residuals.ProbitSpatial}
\alias{residuals.ProbitSpatial}
\title{Extract spatial probit model residuals.}
\usage{
\method{residuals}{ProbitSpatial}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ProbitSpatial}.}

\item{...}{ignored}
}
\value{
Return a vector containing the generalised residuals of the 
	\code{ProbitSpatial} model.
}
\description{
Compute the residuals of an estimated \code{ProbitSpatial} model.
}
