% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitSpatial_R.R
\name{residuals.SpatialProbit}
\alias{residuals.SpatialProbit}
\title{Extract spatial probit model residuals.}
\usage{
\method{residuals}{SpatialProbit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{SpatialProbit}.}

\item{...}{ignored}
}
\value{
Return a vector containing the generalised residuals of the 
	\code{SpatialProbit} model.
}
\description{
Compute the residuals of an estimated \code{SpatialProbit} model.
}

